v1.2.0

+++ OpenBots.Server.Web/Controllers/CredentialsController.cs
@@ -10,6 +10,7 @@ using OpenBots.Server.Model.Attributes;
 using OpenBots.Server.Model.Core;
 using OpenBots.Server.Security;
 using OpenBots.Server.ViewModel;
+using OpenBots.Server.Web.Webhooks;
 using OpenBots.Server.WebAPI.Controllers;
 using System;
 using System.Collections.Generic;
@@ -22,7 +23,7 @@ namespace OpenBots.Server.Web
     /// Controller for credentials
     /// </summary>
     [V1]
-    [Route("api/v{version:apiVersion}/[controller]")]
+    [Route("api/v{apiVersion:apiVersion}/[controller]")]
     [ApiController]
     [Authorize]
     public class CredentialsController : EntityController<Credential>
@@ -32,17 +33,21 @@ namespace OpenBots.Server.Web
         /// </summary>
         ICredentialManager credentialManager;
         private readonly IHttpContextAccessor httpContextAccessor;
+        private readonly IWebhookPublisher webhookPublisher;
+
         public CredentialsController(
             ICredentialRepository repository,
             IMembershipManager membershipManager,
             ApplicationIdentityUserManager userManager,
             ICredentialManager credentialManager,
             IConfiguration configuration,
-            IHttpContextAccessor httpContextAccessor) : base(repository, userManager, httpContextAccessor, membershipManager, configuration)
+            IHttpContextAccessor httpContextAccessor,
+            IWebhookPublisher webhookPublisher) : base(repository, userManager, httpContextAccessor, membershipManager, configuration)
         {
             this.httpContextAccessor = httpContextAccessor;
             this.credentialManager = credentialManager;
             this.credentialManager.SetContext(SecurityContext);
+            this.webhookPublisher = webhookPublisher;
         }
 
         /// <summary>
@@ -124,10 +129,10 @@ namespace OpenBots.Server.Web
             {
                 Guid entityId = new Guid(id);
 
-                var existingAgent = repository.GetOne(entityId);
-                if (existingAgent == null) return NotFound();
+                var existingCredential = repository.GetOne(entityId);
+                if (existingCredential == null) return NotFound();
 
-                if (credentialManager.ValidateRetrievalDate(existingAgent))
+                if (credentialManager.ValidateRetrievalDate(existingCredential))
                 {
                     return await base.GetEntity(id);
                 }
@@ -181,6 +186,7 @@ namespace OpenBots.Server.Web
         /// <response code="400">Bad request, if credential id is not in proper format or a proper Guid</response>
         /// <response code="403">Forbidden</response>
         /// <response code="404">Not found, when no credential exists for the given credential id</response>
+        /// <response code="422">Unprocessable entity</response>
         /// <returns>Credential view details for the given id</returns>
         [HttpGet("password/{id}")]
         [ProducesResponseType(typeof(string), StatusCodes.Status200OK)]
@@ -241,10 +247,6 @@ namespace OpenBots.Server.Web
                 return BadRequest(ModelState);
             }
 
-            Guid entityId = Guid.NewGuid();
-            if (request.Id == null || request.Id.HasValue || request.Id.Equals(Guid.Empty))
-                request.Id = entityId;
-
             try
             {
                 applicationUser = userManager.GetUserAsync(httpContextAccessor.HttpContext.User).Result;
@@ -260,7 +262,16 @@ namespace OpenBots.Server.Web
                     ModelState.AddModelError("Credential", "Credential Name Already Exists");
                     return BadRequest(ModelState);
                 }
-                return await base.PostEntity(request);
+
+                if (!credentialManager.ValidateStartAndEndDates(request))
+                {
+                    ModelState.AddModelError("Credential", "Start and End Date are not valid");
+                    return BadRequest(ModelState);
+                }
+
+                var result = await base.PostEntity(request);
+                await webhookPublisher.PublishAsync("Credentials.NewCredentialCreated", request.Id.ToString(), request.Name).ConfigureAwait(false);
+                return result;
             }
             catch (Exception ex)
             {
@@ -296,12 +307,22 @@ namespace OpenBots.Server.Web
                 Guid entityId = new Guid(id);
 
                 var existingCredential = repository.GetOne(entityId);
-                if (existingCredential == null) return NotFound();
+                if (existingCredential == null)
+                {
+                    ModelState.AddModelError("Credential", "Credential cannot be found or does not exist.");
+                    return NotFound(ModelState);
+                }
 
                 var credential = repository.Find(null, d => d.Name.ToLower(null) == request.Name.ToLower(null) && d.Id != entityId)?.Items?.FirstOrDefault();
                 if (credential != null && credential.Id != entityId)
                 {
                     ModelState.AddModelError("Credential", "Credential Name Already Exists");
+                    return BadRequest(ModelState);
+                }
+
+                if (!credentialManager.ValidateStartAndEndDates(request))
+                {
+                    ModelState.AddModelError("Credential", "Start and End Date are not valid");
                     return BadRequest(ModelState);
                 }
 
@@ -322,6 +343,7 @@ namespace OpenBots.Server.Web
                     existingCredential.PasswordHash = userManager.PasswordHasher.HashPassword(applicationUser, request.PasswordSecret);
                 }
 
+                await webhookPublisher.PublishAsync("Credentials.CredentialUpdated", existingCredential.Id.ToString(), existingCredential.Name).ConfigureAwait(false);
                 return await base.PutEntity(id, existingCredential);
             }
             catch (Exception ex)
@@ -347,6 +369,14 @@ namespace OpenBots.Server.Web
         [ProducesDefaultResponseType]
         public async Task<IActionResult> Delete(string id)
         {
+            var existingCredential = repository.GetOne(Guid.Parse(id));
+            if (existingCredential == null)
+            {
+                ModelState.AddModelError("Credential", "Credential cannot be found or does not exist.");
+                return NotFound(ModelState);
+            }
+
+            await webhookPublisher.PublishAsync("Credentials.CredentialDeleted", existingCredential.Id.ToString(), existingCredential.Name).ConfigureAwait(false);
             return await base.DeleteEntity(id);
         }
 
@@ -370,8 +400,16 @@ namespace OpenBots.Server.Web
             [FromBody] JsonPatchDocument<Credential> request)
         {
             Guid entityId = new Guid(id);
+            var existingCredential = repository.GetOne(entityId);
+            if (existingCredential == null)
+            {
+                ModelState.AddModelError("Credential", "Credential cannot be found or does not exist.");
+                return NotFound(ModelState);
+            }
+
             for (int i = 0; i < request.Operations.Count; i++)
             {
+                // Verify that Credential name is not taken
                 if (request.Operations[i].op.ToString().ToLower() == "replace" && request.Operations[i].path.ToString().ToLower() == "/name")
                 {
                     var credential = repository.Find(null, d => d.Name.ToLower(null) == request.Operations[i].value.ToString().ToLower(null) && d.Id != entityId)?.Items?.FirstOrDefault();
@@ -381,15 +419,37 @@ namespace OpenBots.Server.Web
                         return BadRequest(ModelState);
                     }
                 }
-                    
-                if (request.Operations[i].op.ToString().ToLower() == "replace" && request.Operations[i].path.ToString().ToLower() == "/passwordsecret")
+                   
+                // Generate new password hash
+                if (request.Operations[i].op.ToString().ToLower() == "replace" && request.Operations[i].path.ToString().ToLower() == "/passwordsecret" )
                 {
                     applicationUser = userManager.GetUserAsync(httpContextAccessor.HttpContext.User).Result;
 
                     var passwordHash = userManager.PasswordHasher.HashPassword(applicationUser, request.Operations[i].value.ToString());
                     request.Replace(e => e.PasswordHash, passwordHash);
                 }
+
+                // Verify start-end date range
+                if (request.Operations[i].op.ToString().ToLower() == "replace" && request.Operations[i].path.ToString().ToLower() == "/startdate" 
+                    | request.Operations[i].path.ToString().ToLower() == "/enddate")
+                {
+                    if (request.Operations[i].path.ToString().ToLower() == "/startdate")
+                    {
+                        existingCredential.StartDate = Convert.ToDateTime(request.Operations[i].value.ToString());
+                    }
+                    else
+                    {
+                        existingCredential.EndDate = Convert.ToDateTime(request.Operations[i].value.ToString());
+                    }
+
+                    if (!credentialManager.ValidateStartAndEndDates(existingCredential))
+                    {
+                        ModelState.AddModelError("Credential", "Start and End Date are not valid. End Date must be after the Start Date");
+                        return BadRequest(ModelState);
+                    }
+                }
             }
+            await webhookPublisher.PublishAsync("Credentials.CredentialUpdated", existingCredential.Id.ToString(), existingCredential.Name).ConfigureAwait(false);
             return await base.PatchEntity(id, request);
         }
